Fixed integer overflow in TMVAInterface, a maximum of 2^31 events was al little bit too strict. Now its 2^64 enough for everyone
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@12328 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ analysis/TMVAInterface/include/Teacher.h
@@ -81,13 +81,13 @@ namespace Belle2 {
        * @param prepareOption options which are passed to the TMVA::Factory::PrepareTrainingAndTestTree, in most cases default options should be fine.
        * @param maxEventsPerClass maximum number of events given to TMVA per class. TMVA internally uses a vector instead of a tree and therefore looses out-of-core ability.
        */
-      void train(std::string factoryOption = "!V:!Silent:Color:DrawProgressBar:AnalysisType=Classification", std::string prepareOption = "SplitMode=random:!V", unsigned int maxEventsPerClass = 0);
+      void train(std::string factoryOption = "!V:!Silent:Color:DrawProgressBar:AnalysisType=Classification", std::string prepareOption = "SplitMode=random:!V", unsigned long int maxEventsPerClass = 0);
 
     private:
       /**
        * Train a class against the rest and return ptree with the configuration of the resulting trainig.
        */
-      boost::property_tree::ptree trainClass(std::string factoryOption, std::string prepareOption, std::map<int, unsigned int>& cluster_count, unsigned int maxEventsPerClass, int signalClass);
+      boost::property_tree::ptree trainClass(std::string factoryOption, std::string prepareOption, std::map<int, unsigned long int>& cluster_count, unsigned long int maxEventsPerClass, int signalClass);
 
     private:
       std::string m_prefix; /**< used to identify the outputted training files weights/$prefix_$method.class.C and weights/$prefix_$method.weights.xml */
